+++ src/dandi/types.rs
@@ -1,4 +1,4 @@
-use super::{DandisetId, VersionId};
+use super::{DandiError, DandisetId, VersionId};
 use crate::httputil::HttpUrl;
 use crate::paths::{PureDirPath, PurePath};
 use crate::s3::{PrefixedS3Client, S3Entry, S3Folder, S3Location, S3Object};
@@ -128,51 +128,35 @@ impl<'de> Deserialize<'de> for VersionMetadata {
     }
 }
 
-// Item in a `/dandisets/{dandiset_id}/versions/{version_id}/assets/paths/`
-// response
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
-#[serde(from = "RawFolderEntry")]
-pub(crate) enum FolderEntry {
-    Folder(AssetFolder),
-    Asset { path: PurePath, id: String },
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) struct AssetFolder {
     pub(crate) path: PureDirPath,
 }
 
-impl From<RawFolderEntry> for FolderEntry {
-    fn from(entry: RawFolderEntry) -> FolderEntry {
-        if let Some(asset) = entry.asset {
-            FolderEntry::Asset {
-                path: entry.path,
-                id: asset.asset_id,
-            }
-        } else {
-            FolderEntry::Folder(AssetFolder {
-                path: entry.path.to_dir_path(),
-            })
+#[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
+#[serde(tag = "type", content = "resource", rename_all = "lowercase")]
+pub(super) enum RawAtPath {
+    Folder { path: PurePath },
+    Asset(RawAsset),
+}
+
+impl RawAtPath {
+    pub(super) fn try_unraw(
+        self,
+        endpoint: &super::VersionEndpoint<'_>,
+    ) -> Result<AtPath, DandiError> {
+        match self {
+            RawAtPath::Folder { path } => Ok(AtPath::Folder(AssetFolder {
+                path: path.to_dir_path(),
+            })),
+            RawAtPath::Asset(asset) => Ok(AtPath::Asset(asset.try_into_asset(endpoint)?)),
         }
     }
 }
 
-// Raw item in a `/dandisets/{dandiset_id}/versions/{version_id}/assets/paths/`
-// response
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
-struct RawFolderEntry {
-    path: PurePath,
-    asset: Option<RawFolderEntryAsset>,
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
-struct RawFolderEntryAsset {
-    asset_id: String,
-}
-
 #[allow(clippy::large_enum_variant)]
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) enum AtAssetPath {
+pub(super) enum AtPath {
     Folder(AssetFolder),
     Asset(Asset),
 }
@@ -239,10 +223,24 @@ impl ZarrAsset {
             .find_map(|url| S3Location::parse_url(url.as_url()).ok())
     }
 
-    pub(crate) fn make_resource(&self, value: S3Entry) -> DandiResource {
+    pub(super) fn make_resource(&self, value: S3Entry) -> DandiResource {
         match value {
             S3Entry::Folder(folder) => DandiResource::ZarrFolder(self.make_folder(folder)),
             S3Entry::Object(obj) => DandiResource::ZarrEntry(self.make_entry(obj)),
+        }
+    }
+
+    pub(super) fn make_resource_with_s3(
+        &self,
+        value: S3Entry,
+        s3: PrefixedS3Client,
+    ) -> ZarrResource {
+        match value {
+            S3Entry::Folder(folder) => ZarrResource::Folder {
+                folder: self.make_folder(folder),
+                s3,
+            },
+            S3Entry::Object(obj) => ZarrResource::Entry(self.make_entry(obj)),
         }
     }
 
@@ -349,17 +347,6 @@ pub(crate) enum DandiResource {
     ZarrEntry(ZarrEntry),
 }
 
-impl DandiResource {
-    pub(super) fn with_s3(self, s3: PrefixedS3Client) -> DandiResourceWithS3 {
-        match self {
-            DandiResource::Folder(r) => DandiResourceWithS3::Folder(r),
-            DandiResource::Asset(r) => DandiResourceWithS3::Asset(r),
-            DandiResource::ZarrFolder(folder) => DandiResourceWithS3::ZarrFolder { folder, s3 },
-            DandiResource::ZarrEntry(r) => DandiResourceWithS3::ZarrEntry(r),
-        }
-    }
-}
-
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) struct ZarrFolder {
     pub(crate) zarr_path: PurePath,
@@ -404,32 +391,28 @@ pub(crate) struct ZarrEntry {
 }
 
 #[derive(Clone, Debug)]
-pub(super) enum DandiResourceWithS3 {
-    Folder(AssetFolder),
-    Asset(Asset),
-    ZarrFolder {
+pub(super) enum ZarrResource {
+    Folder {
         folder: ZarrFolder,
         s3: PrefixedS3Client,
     },
-    ZarrEntry(ZarrEntry),
+    Entry(ZarrEntry),
 }
 
-impl From<AtAssetPath> for DandiResourceWithS3 {
-    fn from(value: AtAssetPath) -> DandiResourceWithS3 {
+impl From<AtPath> for DandiResource {
+    fn from(value: AtPath) -> DandiResource {
         match value {
-            AtAssetPath::Folder(r) => DandiResourceWithS3::Folder(r),
-            AtAssetPath::Asset(r) => DandiResourceWithS3::Asset(r),
+            AtPath::Folder(r) => DandiResource::Folder(r),
+            AtPath::Asset(r) => DandiResource::Asset(r),
         }
     }
 }
 
-impl From<DandiResourceWithS3> for DandiResource {
-    fn from(value: DandiResourceWithS3) -> DandiResource {
+impl From<ZarrResource> for DandiResource {
+    fn from(value: ZarrResource) -> DandiResource {
         match value {
-            DandiResourceWithS3::Folder(r) => DandiResource::Folder(r),
-            DandiResourceWithS3::Asset(r) => DandiResource::Asset(r),
-            DandiResourceWithS3::ZarrFolder { folder, .. } => DandiResource::ZarrFolder(folder),
-            DandiResourceWithS3::ZarrEntry(r) => DandiResource::ZarrEntry(r),
+            ZarrResource::Folder { folder, .. } => DandiResource::ZarrFolder(folder),
+            ZarrResource::Entry(r) => DandiResource::ZarrEntry(r),
         }
     }
 }