+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+In Development
+--------------
+- Use the Archive API's `/webdav/assets/atpath/` endpoint for fetching
+  information about assets
+
 v0.6.0 (2025-03-06)
 -------------------
 - Serve a `/robots.txt` file denying all robots